+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added KubeJS integration - <WIKI>
 - Added Jade integration to Lightroom: it now shows printing progress arrow. 
 - Added config option to disable attacking while looking through Viewfinder.
+- Added config option to delay capture (useful when GUI elements are not hidden fully, which happens with some shaders)
 - Stacking two Photographs (by right-clicking in GUI) will keep the resulted item in slot instead of picking it up.
 - Camera tooltip now shows exposed/available frames of the inserted Film Roll.
 - Developed rolls now show frame size (if custom) in tooltip, same as undeveloped rolls.